Implemented CommonGELL
The CommonGell is a GameObject containing a special GameEventListenerList that listens to all common gameplay events: Pause, Resume, GameOver, and LevelBegin (more to come)

The script is attached to a GameObject. Drag the CommonGell GO inside any Prefab, and it will listen to those common game events as long as they are implemented in code.

This is done by retrieving any MonoBehaviour within the GameObject that inherits ICommonGameEvents. The methods are placed in corresponding UnityActions using the += operator, and then assigned to their corresponding GameEventListeners.

Important notes:

If a Behaviour script doesn't need to implement a method for pause, game over, etc., it should leave the method empty.

CommonGELL won't work on its own, the designer must ensure its corresponding methods are implemented in the GameObject's Behaviour scripts.

I will implement some warning logs later to deal with bugs.