Fixed level start bug
Bug:

LevelCompleteMenuBehaviour's Awake method was being called after its Hide() method,returning an error because the script never made references to its components before it had to Hide them.

The solution was to put the GameManager's call to set the Game State to LevelStart in a Start method instead of an Awake method. The game can only start once the assets have been assigned.