Cleaning up code
static methods do not like calling private methods, likely because the private methods are (1) not static (and therefore...) (2) tied to an instance of the SwimmerInput class.

This complicates things because I don't know which variables, methods, or classes should be made static now... this made need to be a singleton class.